Strengthen thunk_force_forced by proving that [R] is not required.
One might hope to strengthen this law even further by proving that
the postcondition [□ φ v] can be obtained *if* [n] is zero, or *if*
one pays [n] time credits. However, my attempts to do so have failed,
due to technical difficulties. (In ThunksStep, the fact that the piggy
bank has zero debit does not imply that the underlying thunk has zero
debit. This creates complications.) I abandon this idea.